Added "must add" functionality, fixed more trimming issues, added defaults for write_input templates
Added ability to specify groups that need to be included in the models but are not part of the seed.
Trimming script checks for incomplete prolines/adjacent CAs iteratively after other trimming steps until nothing more changes.
Changed "path to RIN" to "path to scripts" in driver_input to make things clearer.
Added default input templates for each program in write_input.