Template pdb creation separated from write_input, check_duplicate_models rewrite, fixes to trimming procedure
Template pdb files now created by separate script so that they are no longer dependent on starting with a H-only optimisation.
Driver updated to reflect this change.
Rewrote check_duplicate_models to be more robust and not need file with list of names.
Some wonky model numbering fixed in the trimming script and now it will make seed only model as well.